Chunk 10: split GroupRecord.jsx and TeamRecord.jsx (M2)
Extraction only, no behaviour change. Each page already had two top-level
sub-components defined in-file; moved them to their own files:

GroupRecord.jsx (1,128 -> 152): groups/GroupDetails.jsx + groups/GroupLedger.jsx
TeamRecord.jsx  (889 -> 125):  teams/TeamDetails.jsx  + teams/TeamLedger.jsx

The page files now own only tab routing and render the sub-components.
Frontend suite green (53 files, 140 tests); lint 0 errors, Prettier clean.
